requirements.txt
schedule
slackclient>=2.0.0
slackeventsapi>=2.1.0
Flask>=1.1.1
certifi